chore: move and rename files into more logical grouping
