+++ apps/trading/components/ticket-container/elements/slider.tsx
