+++ apps/trading/components/ticket-container/ticket-default/size-slider.tsx
@@ -4,12 +4,12 @@ import { useActiveOrders } from '@vegaprotocol/orders';
 import { useVegaWallet } from '@vegaprotocol/wallet-react';
 import { useOpenVolume } from '@vegaprotocol/positions';
 
-import { Slider } from './slider';
-import { useTicketContext } from './ticket-context';
+import { Slider } from '../elements/slider';
+import { useTicketContext } from '../ticket-context';
 
-import * as defaultUtils from './ticket-default/utils';
-import * as utils from './utils';
-import { useForm } from './use-form';
+import * as defaultUtils from './utils';
+import * as utils from '../utils';
+import { useForm } from '../use-form';
 import { useState } from 'react';
 
 export const SizeSlider = ({ price }: { price: BigNumber | undefined }) => {
