+++ apps/trading/components/ticket-container/ticket-spot/ticket-stop-market.tsx
@@ -11,13 +11,11 @@ import {
 } from '@vegaprotocol/types';
 import { useVegaTransactionStore } from '@vegaprotocol/web3';
 import { mapFormValuesToStopOrdersSubmission } from '@vegaprotocol/deal-ticket';
-import { useMarkPrice } from '@vegaprotocol/markets';
-import { toBigNum } from '@vegaprotocol/utils';
 
 import { FieldControls, Form, FormGrid, FormGridCol } from '../elements/form';
 import { type FormFieldsStopMarket, schemaStopMarket } from '../schemas';
 import { TicketTypeSelect } from '../ticket-type-select';
-import { type FormProps } from '../ticket-default';
+import { type FormProps } from './ticket';
 import { NON_PERSISTENT_TIF_OPTIONS } from '../constants';
 import { TicketEventUpdater } from '../ticket-events';
 import { useTicketContext } from '../ticket-context';
@@ -25,8 +23,6 @@ import { SubmitButton } from '../elements/submit-button';
 import { useT } from '../../../lib/use-t';
 
 import * as Fields from '../fields';
-
-import { SizeSliderStop } from '../size-slider-stop';
 
 export const TicketStopMarket = (props: FormProps) => {
   const t = useT();
@@ -63,12 +59,6 @@ export const TicketStopMarket = (props: FormProps) => {
   const isPersistent = !NON_PERSISTENT_TIF_OPTIONS.includes(tif);
   const oco = form.watch('oco');
 
-  const { data: markPrice } = useMarkPrice(ticket.market.id);
-  const price =
-    markPrice && markPrice !== null
-      ? toBigNum(markPrice, ticket.market.decimalPlaces)
-      : undefined;
-
   return (
     <FormProvider {...form}>
       <TicketEventUpdater />
@@ -101,7 +91,6 @@ export const TicketStopMarket = (props: FormProps) => {
           </FieldControls>
           <Fields.StopSize />
         </div>
-        <SizeSliderStop price={price} />
         <FormGrid>
           <FormGridCol>
             {isPersistent ? <Fields.PostOnly /> : <Fields.ReduceOnly />}
